+++ packages/api/lib/controllers/usersController.js
@@ -43,7 +43,7 @@ const { User } = require('@sensebox/opensensemap-api-models'),
  * @apiSuccess (Created 201) {String} refreshToken valid refresh token
  * @apiSuccess (Created 201) {Object} data `{ "user": {"name":"fullname","email":"test@test.de","role":"user","language":"en_US","boxes":[],"emailIsConfirmed":false} }`
  */
-const registerUser = async function registerUser (req, res, next) {
+const registerUser = async function registerUser (req, res) {
   const { email, password, language, name } = req._userParams;
 
   try {
@@ -56,10 +56,10 @@ const registerUser = async function registerUser (req, res, next) {
 
       return res.send(201, { code: 'Created', message: 'Successfully registered new user', data: { user: newUser }, token, refreshToken });
     } catch (err) {
-      return next(new InternalServerError(`User successfully created but unable to create jwt token: ${err.message}`));
+      return Promise.reject(new InternalServerError(`User successfully created but unable to create jwt token: ${err.message}`));
     }
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -77,7 +77,7 @@ const registerUser = async function registerUser (req, res, next) {
  * @apiSuccess {Object} data `{ "user": {"name":"fullname","email":"test@test.de","role":"user","language":"en_US","boxes":[],"emailIsConfirmed":false} }`
  * @apiError {String} 403 Unauthorized
  */
-const signIn = async function signIn (req, res, next) {
+const signIn = async function signIn (req, res) {
   const { email: emailOrName, password } = req._userParams;
 
   try {
@@ -87,7 +87,7 @@ const signIn = async function signIn (req, res, next) {
       .exec();
 
     if (!user) {
-      throw new ForbiddenError('User and or password not valid!');
+      return Promise.reject(new ForbiddenError('User and or password not valid!'));
     }
 
     if (await user.checkPassword(password)) {
@@ -97,9 +97,10 @@ const signIn = async function signIn (req, res, next) {
     }
   } catch (err) {
     if (err.name === 'ModelError' && err.message === 'Password incorrect') {
-      return handleError(new ForbiddenError('User and or password not valid!'), next);
+      return handleError(new ForbiddenError('User and or password not valid!'));
     }
-    handleError(err, next);
+
+    return handleError(err);
   }
 };
 
@@ -116,12 +117,12 @@ const signIn = async function signIn (req, res, next) {
  * @apiSuccess {Object} data `{ "user": {"name":"fullname","email":"test@test.de","role":"user","language":"en_US","boxes":[],"emailIsConfirmed":false} }`
  * @apiError {Object} Forbidden `{"code":"ForbiddenError","message":"Refresh token invalid or too old. Please sign in with your username and password."}`
  */
-const refreshJWT = async function refreshJWT (req, res, next) {
+const refreshJWT = async function refreshJWT (req, res) {
   try {
     const { token, refreshToken, user } = await refreshJwt(req._userParams.token);
     res.send(200, { code: 'Authorized', message: 'Successfully refreshed auth', data: { user }, token, refreshToken });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -134,7 +135,7 @@ const refreshJWT = async function refreshJWT (req, res, next) {
  * @apiSuccess {String} code `Ok`
  * @apiSuccess {String} message `Successfully signed out`
  */
-const signOut = function signOut (req, res) {
+const signOut = async function signOut (req, res) {
   invalidateToken(req);
 
   return res.send(200, { code: 'Ok', message: 'Successfully signed out' });
@@ -150,12 +151,12 @@ const signOut = function signOut (req, res) {
  * @apiSuccess {String} message `Password reset initiated`
  */
 // generate new password reset token and send the token to the user
-const requestResetPassword = async function requestResetPassword (req, res, next) {
+const requestResetPassword = async function requestResetPassword (req, res) {
   try {
     await User.initPasswordReset(req._userParams);
     res.send(200, { code: 'Ok', message: 'Password reset initiated' });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -170,12 +171,12 @@ const requestResetPassword = async function requestResetPassword (req, res, next
  * @apiSuccess {String} message `Password successfully changed. You can now login with your new password`
  */
 // set new password with reset token as auth
-const resetPassword = async function resetPassword (req, res, next) {
+const resetPassword = async function resetPassword (req, res) {
   try {
     await User.resetPassword(req._userParams);
     res.send(200, { code: 'Ok', message: 'Password successfully changed. You can now login with your new password' });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -189,29 +190,30 @@ const resetPassword = async function resetPassword (req, res, next) {
  * @apiSuccess {String} code `Ok`
  * @apiSuccess {String} message `E-Mail successfully confirmed. Thank you`
  */
-const confirmEmailAddress = async function confirmEmailAddress (req, res, next) {
+const confirmEmailAddress = async function confirmEmailAddress (req, res) {
   try {
     await User.confirmEmail(req._userParams);
     res.send(200, { code: 'Ok', message: 'E-Mail successfully confirmed. Thank you' });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
 /**
- * @api {post} /users/me/boxes list all boxes of the signed in user
+ * @api {post} /users/me/boxes list all boxes and sharedBoxes of the signed in user
  * @apiName getUserBoxes
- * @apiDescription List all boxes of the signed in user with secret fields
+ * @apiDescription List all boxes and sharedBoxes of the signed in user with secret fields
  * @apiGroup Users
  * @apiSuccess {String} code `Ok`
  * @apiSuccess {String} data A json object with a single `boxes` array field
  */
-const getUserBoxes = async function getUserBoxes (req, res, next) {
+const getUserBoxes = async function getUserBoxes (req, res) {
   try {
     const boxes = await req.user.getBoxes();
-    res.send(200, { code: 'Ok', data: { boxes } });
+    const sharedBoxes = await req.user.getSharedBoxes();
+    res.send(200, { code: 'Ok', data: { boxes: boxes, sharedBoxes: sharedBoxes } });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -222,7 +224,7 @@ const getUserBoxes = async function getUserBoxes (req, res, next) {
  * @apiGroup Users
  * @apiUse JWTokenAuth
  */
-const getUser = function getUser (req, res) {
+const getUser = async function getUser (req, res) {
   res.send(200, { code: 'Ok', data: { me: req.user } });
 };
 
@@ -238,7 +240,7 @@ const getUser = function getUser (req, res) {
  * @apiParam {String} [newPassword] the new password for this user. Should be at least 8 characters long.
  * @apiParam {String} currentPassword the current password for this user.
  */
-const updateUser = async function updateUser (req, res, next) {
+const updateUser = async function updateUser (req, res) {
   try {
     const { updated, signOut, messages, updatedUser } = await req.user.updateUser(req._userParams);
     if (updated === false) {
@@ -250,7 +252,7 @@ const updateUser = async function updateUser (req, res, next) {
     }
     res.send(200, { code: 'Ok', message: `User successfully saved.${messages.join('.')}`, data: { me: updatedUser } });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -263,7 +265,7 @@ const updateUser = async function updateUser (req, res, next) {
  * @apiParam {String} password the current password for this user.
  */
 
-const deleteUser = async function deleteUser (req, res, next) {
+const deleteUser = async function deleteUser (req, res) {
   const { password } = req._userParams;
 
   try {
@@ -275,7 +277,7 @@ const deleteUser = async function deleteUser (req, res, next) {
     clearCache(['getBoxes', 'getStats']);
     postToMattermost(`User deleted: ${req.user.name} (${redactEmail(req.user.email)})`);
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };
 
@@ -288,7 +290,7 @@ const deleteUser = async function deleteUser (req, res, next) {
  * @apiSuccess {String} code `Ok`
  * @apiSuccess {String} message `Email confirmation has been sent to <emailaddress>`
  */
-const requestEmailConfirmation = async function requestEmailConfirmation (req, res, next) {
+const requestEmailConfirmation = async function requestEmailConfirmation (req, res) {
   try {
     const result = await req.user.resendEmailConfirmation();
     let usedAddress = result.email;
@@ -297,7 +299,7 @@ const requestEmailConfirmation = async function requestEmailConfirmation (req, r
     }
     res.send(200, { code: 'Ok', message: `Email confirmation has been sent to ${usedAddress}` });
   } catch (err) {
-    handleError(err, next);
+    return handleError(err);
   }
 };