+++ packages/api/lib/routes.js
@@ -26,7 +26,7 @@ const spaces = function spaces (num) {
  * @apiDescription Returns all routes of this API in human readable format
  * @apiGroup Misc
  */
-const printRoutes = function printRoutes (req, res) {
+const printRoutes = async function printRoutes (req, res) {
   res.header('Content-Type', 'text/plain; charset=utf-8');
 
   const lines = [
@@ -78,6 +78,7 @@ const routes = {
     { path: `${statisticsPath}/descriptive`, method: 'get', handler: statisticsController.descriptiveStatisticsHandler, reference: 'api-Statistics-descriptive' },
     { path: `${boxesPath}`, method: 'get', handler: boxesController.getBoxes, reference: 'api-Boxes-getBoxes' },
     { path: `${boxesPath}/data`, method: 'get', handler: measurementsController.getDataMulti, reference: 'api-Measurements-getDataMulti' },
+    { path: `${boxesPath}/data/bytag`, method: 'get', handler: measurementsController.getDataByGroupTag, reference: 'api-Measurements-getDataByGroupTag' },
     { path: `${boxesPath}/:boxId`, method: 'get', handler: boxesController.getBox, reference: 'api-Boxes-getBox' },
     { path: `${boxesPath}/:boxId/sensors`, method: 'get', handler: measurementsController.getLatestMeasurements, reference: 'api-Measurements-getLatestMeasurements' },
     { path: `${boxesPath}/:boxId/sensors/:sensorId`, method: 'get', handler: measurementsController.getLatestMeasurements, reference: 'api-Measurements-getLatestMeasurementOfSensor' },
@@ -116,13 +117,11 @@ const routes = {
     { path: `${managementPath}/boxes/:boxId`, method: 'get', handler: managementController.getBox, reference: 'api-Admin-getBox' },
     { path: `${managementPath}/boxes/:boxId`, method: 'put', handler: managementController.updateBox, reference: 'api-Admin-updateBox' },
     { path: `${managementPath}/boxes/delete`, method: 'post', handler: managementController.deleteBoxes, reference: 'api-Admin-deleteBoxes' },
-
     { path: `${managementPath}/users`, method: 'get', handler: managementController.listUsers, reference: 'api-Admin-listUsers' },
     { path: `${managementPath}/users/:userId`, method: 'get', handler: managementController.getUser, reference: 'api-Admin-getUser' },
     { path: `${managementPath}/users/:userId`, method: 'put', handler: managementController.updateUser, reference: 'api-Admin-updateUser' },
     { path: `${managementPath}/users/delete`, method: 'post', handler: managementController.deleteUsers, reference: 'api-Admin-deleteUsers' },
     { path: `${managementPath}/users/:userId/exec`, method: 'post', handler: managementController.execUserAction, reference: 'api-Admin-execUserAction' },
-
   ]
 };
 
@@ -136,16 +135,21 @@ const initRoutes = function initRoutes (server) {
   }
 
   // Attach secured routes (needs authorization through jwt)
-  server.use(verifyJwt);
-
+  // The .use() method runs now for all routes
+  // https://github.com/restify/node-restify/issues/1685
   for (const route of routes.auth) {
-    server[route.method]({ path: route.path }, route.handler);
+    server[route.method]({ path: route.path }, [verifyJwt, route.handler]);
   }
 
-  server.use(checkPrivilege);
-
+  // Attach verifyJwt and checkPrivilage routes (needs authorization through jwt)
+  // The .use() method runs now for all routes
+  // https://github.com/restify/node-restify/issues/1685
   for (const route of routes.management) {
-    server[route.method]({ path: route.path }, route.handler);
+    server[route.method]({ path: route.path }, [
+      verifyJwt,
+      checkPrivilege,
+      route.handler,
+    ]);
   }
 };