+++ tests/tests/002-location_tests.js
@@ -284,6 +284,14 @@ describe('openSenseMap API locations tests', function () {
         });
     });
 
+    it('should reject filtering boxes near a location with wrong parameter values', function () {
+      return chakram.get(`${BASE_URL}?near=test,60`).then(function (response) {
+        expect(response).to.have.status(422);
+
+        return chakram.wait();
+      });
+    });
+
   });
 
   describe('POST /boxes/:boxID/:sensorID', function () {